[#76] refactor: GCP Secret-Manager 클라이언트를 빈으로 관리
리뷰 반영
